New TeleOp Main
Additional Cleanup as well
