+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/testMain.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.MainCode.TeleOp;
+package org.firstinspires.ftc.teamcode.testing;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
